refactor: move UI helpers
